+++ packages/orq-rc/src/models/components/index.ts
@@ -415,6 +415,7 @@ export * from "./projectbudgetscoperestresponse.js";
 export * from "./projectscope.js";
 export * from "./providerbudgetscope.js";
 export * from "./providertoolinput.js";
+export * from "./publiccompactionitem.js";
 export * from "./publiccontact.js";
 export * from "./publicembeddingdata.js";
 export * from "./publicembeddingusage.js";