+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3890,7 +3890,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01M37EHKWSJ20RC95J5NTFEAYS"),
+  id: z.string().default("01M3BK2D94V8AC9F40V5F7PMPG"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),