+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3949,7 +3949,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01M37EHKXH65XM6D9BYRP33DG9"),
+  id: z.string().default("01M3BK2D9Q5PXKFC8YYNZ9SZHG"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>