+++ packages/orq-rc/RELEASES.md
@@ -12609,3 +12609,13 @@ Based on:
 - [typescript v4.16.0-rc.16] packages/orq-rc
 ### Releases
 - [NPM v4.16.0-rc.16] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.16 - packages/orq-rc
+
+## 2026-09-25 06:14:59
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.799.0 (2.941.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.16.0-rc.17] packages/orq-rc
+### Releases
+- [NPM v4.16.0-rc.17] https://www.npmjs.com/package/@orq-ai/node/v/4.16.0-rc.17 - packages/orq-rc